fix: only force server id generation if requested
An inverted boolean expression accidentally always forced the
driver to force the server to generate document ids.

NODE-2923
